refactor: extract file-scanning into files.rs, add version-mismatch test
- Move template_files_in_dir, collect_template_files, and is_template_file
  from loader.rs into a new files.rs module; re-export is_template_file
  from loader.rs so existing callers (doctor.rs) are unaffected
- cache.rs now imports template_files_in_dir from super::files, breaking
  the loader↔cache circular module dependency
- Add version_mismatch_returns_none test for the CACHE_VERSION check path
