+++ src/template/loader.rs
@@ -7,9 +7,13 @@ use anyhow::{Context, Result};
 use crate::config;
 
 use super::catalog::{TemplateCatalog, TemplateCatalogEntry, TemplateScope, TemplateSource};
+use super::files::template_files_in_dir;
 use super::parser::parse_template_hcl;
 use super::schema::TemplateFile;
 use super::validator::validate_template_file;
+
+// Re-export for callers that import is_template_file from this module (e.g. doctor.rs).
+pub use super::files::is_template_file;
 
 pub fn load_catalog(cwd: &Path) -> Result<TemplateCatalog> {
     let global_dir = config::global_templates_dir();
@@ -120,49 +124,6 @@ fn load_file_into_catalog(
     }
 
     Ok(())
-}
-
-pub(super) fn template_files_in_dir(dir: &Path) -> Result<Vec<PathBuf>> {
-    if !dir.exists() {
-        return Ok(Vec::new());
-    }
-    let dir = dir.canonicalize().with_context(|| {
-        format!(
-            "failed to canonicalize template directory {}",
-            dir.display()
-        )
-    })?;
-    let mut files = Vec::new();
-    collect_template_files(&dir, &mut files)?;
-    files.sort();
-    Ok(files)
-}
-
-fn collect_template_files(dir: &Path, files: &mut Vec<PathBuf>) -> Result<()> {
-    for entry in fs::read_dir(dir)
-        .with_context(|| format!("failed listing template directory {}", dir.display()))?
-    {
-        let entry = entry?;
-        let path = entry.path();
-        let file_type = entry
-            .file_type()
-            .with_context(|| format!("failed inspecting template path {}", path.display()))?;
-        if file_type.is_dir() {
-            collect_template_files(&path, files)?;
-            continue;
-        }
-        if file_type.is_file() && is_template_file(&path) {
-            files.push(path);
-        }
-    }
-    Ok(())
-}
-
-pub fn is_template_file(path: &Path) -> bool {
-    path.extension()
-        .and_then(|s| s.to_str())
-        .map(|s| s == "hcl")
-        .unwrap_or(false)
 }
 
 #[cfg(test)]
