+++ src/template/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod cache;
 pub mod catalog;
+pub mod files;
 pub mod import;
 pub mod loader;
 pub mod parser;
